requirements.txt
sentence-transformers
pdfplumber
langchain
chromadb
transformers
torch
pyyaml
python-dotenv
langchain-community
pandas
